Add support for allowlisting hosts of images.
This commit allows us to permit certain domain hosts to use our service
so that it does not get abused by third parties. In the future I would
like to add * support, e.g. *.monstercat.com.
